Simplify output format
This version of roll does away with a lot of output ceremony, opting
to display only comments, expressions, and total results. In order to
get individual roll values, it will be necessary to input your rolls
individually into a given formula--but overall I expect this version
to be much easier to read at a glance.